Got rid of the "Item" class as it was a link to the sql database class, Instead I now use the database class renamed to "ItemDb". Added a Struct that is used to hold the items data from a database row when the "ReadRow" function is called. This struct could be seen as a replacement for the item class, instead of linking tot he database though, it just holds data passed from the database.
Implemented the sm-2 algorithm into the super memo function. The program now calculates everything correctly, however it doesn't use the data to properly display the kanji not learnt.

Also added two new columns to the database, "LastDate" and "Learnt" which contain the date for when the kanji was last accessed (last reviewed) and the date for when the kanji was originally learnt respectively.